WIP add pattern tests
